requirements.txt
google-api-python-client==2.133.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
Flask==3.0.3

